Refactor package structure for test module
