+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/controller/ComplexSampleController.kt
@@ -1,4 +1,4 @@
-package io.github.bgmsound.documentify.sample.mvc
+package io.github.bgmsound.documentify.sample.mvc.controller
 
 import io.github.bgmsound.documentify.sample.mvc.dto.response.ComplexSampleResponse
 import io.github.bgmsound.documentify.sample.mvc.dto.response.NestedSampleResponse
