+++ documentify-sample/mvc-sample/src/test/kotlin/io/github/bgmsound/documentify/sample/mvc/documentation/ErrorDocs.kt
@@ -2,7 +2,7 @@ package io.github.bgmsound.documentify.sample.mvc.documentation
 
 import io.github.bgmsound.documentify.core.specification.schema.Method
 import io.github.bgmsound.documentify.mvc.Documentify
-import io.github.bgmsound.documentify.sample.mvc.ErrorController
+import io.github.bgmsound.documentify.sample.mvc.controller.ErrorController
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.restdocs.RestDocumentationContextProvider
