+++ documentify-sample/mvc-sample/src/test/kotlin/io/github/bgmsound/documentify/sample/mvc/documentation/NestedSampleDocs.kt
@@ -2,7 +2,7 @@ package io.github.bgmsound.documentify.sample.mvc.documentation
 
 import io.github.bgmsound.documentify.core.specification.schema.Method
 import io.github.bgmsound.documentify.mvc.Documentify
-import io.github.bgmsound.documentify.sample.mvc.NestedSampleController
+import io.github.bgmsound.documentify.sample.mvc.controller.NestedSampleController
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.restdocs.RestDocumentationContextProvider
