+++ documentify-sample/reactive-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/reactive/controller/SampleController.kt
@@ -1,9 +1,17 @@
-package io.github.bgmsound.documentify.sample.reactive
+package io.github.bgmsound.documentify.sample.reactive.controller
 
 import io.github.bgmsound.documentify.sample.reactive.dto.request.SampleRequest
 import io.github.bgmsound.documentify.sample.reactive.dto.response.SampleResponse
 import kotlinx.coroutines.reactive.awaitFirst
-import org.springframework.web.bind.annotation.*
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PatchMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.bind.annotation.RestController
 import reactor.core.publisher.Mono
 
 @RestController
