+++ documentify-sample/reactive-sample/src/test/kotlin/io/github/bgmsound/documentify/sample/reactive/documentation/SnakeSampleDocs.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import io.github.bgmsound.documentify.core.specification.schema.Method
 import io.github.bgmsound.documentify.reactive.Documentify
-import io.github.bgmsound.documentify.sample.reactive.SampleController
+import io.github.bgmsound.documentify.sample.reactive.controller.SampleController
 import kotlinx.coroutines.test.runTest
 
 import org.junit.jupiter.api.BeforeEach
